Revert "Se agrega pantalla de inicio se sesion, pantalla de registro y listad…"
