--- app/components/List/styles.js
@@ -1,9 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  container: {
-    // padding: 20,
-  },
-});
-
-export default styles;
