--- app/containers/Batches/styles.js
@@ -1,11 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  container: {
-    paddingTop: 15,
-    paddingLeft: 15,
-    paddingRight: 15,
-  },
-});
-
-export default styles;
