+++ app/containers/Login/index.js
@@ -1,11 +1,26 @@
 import * as React from 'react';
-import { View } from "react-native";
+import { StyleSheet, View } from "react-native";
 import { Text } from 'react-native-elements';
 
-import Button from '../../components/Button';
-import InputLogin from '../../components/inputs/InputLogin';
+import EntryButton from '../../components/buttons/EntryButton';
 import Margin from '../../components/Margin';
-import styles from './styles';
+
+import * as theme from '../../theme';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 20,
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    backgroundColor: theme.backgroundColor,
+  },
+  text: {
+    color: theme.brandColor000,
+    textAlign: 'center',
+  }
+});
 
 const Login = ({ navigation }) => {
   return (
@@ -22,33 +37,25 @@ const Login = ({ navigation }) => {
         </Text>
       </Margin>
 
-      <InputLogin
-        placeholder='Ingrese email'
+      <EntryButton
+        title="Entrar"
+        onPress={() =>
+          navigation.navigate('Private', { name: 'Jane' })
+        }
       />
-
-      <InputLogin
-        placeholder='Ingrese contraseña'
-        secureTextEntry={true}
-      />
-      
-      <Margin size={36}>
-        <Button
-          title="Entrar"
-          onPress={() => navigation.navigate('Private')}
-        />
-      </Margin>
 
       <Margin size={14}>
         <Text style={styles.text}>
           Tambien puedes
         </Text>
-        
-        <Button
-          title='Registro'
-          type='clear'
-          onPress={() => navigation.navigate('SignIn')}
-        />
       </Margin>
+
+      <EntryButton
+        title="Registro"
+        onPress={() =>
+          navigation.navigate('SignIn', { name: 'Jane' })
+        }
+      />
     </View>
   );
 };
